Implementation of template token insertion buttons
The previously static listings of template tokens have been replaced with a special button component that allows you to insert template tokens into the template string by clicking. This lets you build template strings quicker when setting up new rules or workspaces.

The insertion happens at the end of the template string if the user has no selection in the template string input at the time of the button click; otherwise the editor inserts at the user's selection in the template string. A ranged selection replaces the range with the template token and places the cursor after the insertion. (I wanted to maintain the range, but the API doesn't let me do this.) Inserting this way does not break focus from the input. Inserting without focus does not steal focus to the editor.

That said, two things need clarification:
1. the visual design of the template buttons. They should be recognizable as buttons, but the basic grey design may not be desirable or work too well
2. Should the editor always steal focus to typing, even if you insert a token without having set focus into the editor beforehand? You'd probably want to insert spaces etc between tokens, and right now you need to set focus manually. I am unsure what's a more predictable user experience.