Simplify components because inputs are smart
Turns out input components are stateful - they themselves remember where focus was set. That also means it's entirely unneeded to save input before the focus is stolen away from the component. Just run the main callback, and have it ask the input component what its last selection state was (even if the focus isn't there). The behavior is the exact same as previously.

This cleans up the code a bunch and removes redundant code on my end.