feat: make buttons subtle
The normal buttons are very "fat" in the visual impact. The use of the "subtle" class makes them not look like buttons on default and restores the text styling that was lost on the component change.

This hides their interactive behavior somewhat strongly, but you can still discover that they are buttons if you hover over them to copy the tokens (which at least I tried doing previously) so this feature is probably discoverable enough right now.

Otherwise, see the exp/template-button-styling branch.